// src/features/editableProfileCard/model/types/editableProfileCardSchema.ts
import { ValidateProfileError } from '../constants/constants';

import { Profile } from '@/entities/Profile';

export interface ProfileSchema {
    data?: Profile;
    form?: Profile;
    isLoading?: boolean;
    error?: string;
    readonly?: boolean;
    validateErrors?: ValidateProfileError[];
}